Custom data structure & iterator, cmake stuff
holy crap, we're not just using standard stuff now!
we now use an unordered_map<string, Key *>, where Keys are
linked lists containing possible values, with their weight
(so we can algorithmerize the crap out of this later).

I also implemented an iterator for all of this. Boy, that took ages.

We now also have cmake support, because that's probably
going to start being required at some point and it's better
than a Makefile calling clang++...